Removed mock-flavor version of SchedulerProvider, instead returning in-memory database
When running an E2E test with Espresso in the mock flavor, the mock version
of SchedulerProvider provides the Trampoline scheduler for both IO and
Android's Main Thread schedulers - this causes Room to thrown an error (duh).

The downside to this is, I have to manually initialize Mockito's mock of
SchedulerProvider. Minor incontinence to ensure my integration/E2E tests
are using an in-memory database.